fix race in TestQmux
Adds synchronization to wait for the test to complete before closing
the listener Session. Previously `sess.Open` would return immediately,
allowing the test code to proceed to reading from the channel sooner.
Now that it waits for the `Accept`, the `Accept`-side was writing and
then closing the session before the data was read.