session: timeout for calling Accept on new channel (#30)
Adds a timeout for calling Accept when a request to open a new channel is
received. This should help detect bugs with missing the Respond() go routine
to handle incoming channels.

Also fixes a race condition in TestQmux by adds synchronization to wait
for the test to complete before closing the listener Session. Previously
`sess.Open` would return immediately, allowing the test code to proceed
to reading from the channel sooner. Now that it waits for the `Accept`,
the `Accept`-side was writing and then closing the session before the
data was read.